# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Archivista API",
        "title": "Archivista API",
        "contact": {
            "name": "Archivista Contributors",
            "url": "https://github.com/in-toto/archivista/issues/new"
        },
        "license": {
            "url": "https://opensource.org/licenses/Apache-2"
        },
        "version": "v1"
    },
    "paths": {
        "/download/{gitoid}": {
            "get": {
                "description": "download an attestation",
                "produces": [
                    "application/json"
                ],
                "summary": "Download",
                "deprecated": true,
                "parameters": [
                    {
                        "type": "string",
                        "description": "gitoid",
                        "name": "gitoid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dsse.Envelope"
                        }
                    }
                }
            }
        },
        "/upload": {
            "post": {
                "description": "stores an attestation",
                "produces": [
                    "application/json"
                ],
                "summary": "Upload",
                "deprecated": true,
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.StoreResponse"
                        }
                    }
                }
            }
        },
        "/v1/download/{gitoid}": {
            "get": {
                "description": "download an attestation",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "attestation"
                ],
                "summary": "Download",
                "parameters": [
                    {
                        "type": "string",
                        "description": "gitoid",
                        "name": "gitoid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dsse.Envelope"
                        }
                    }
                }
            }
        },
        "/v1/query": {
            "post": {
                "description": "GraphQL query",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "graphql"
                ],
                "summary": "Query GraphQL",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/archivista.Resolver"
                        }
                    }
                }
            }
        },
        "/v1/upload": {
            "post": {
                "description": "stores an attestation",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "attestation"
                ],
                "summary": "Upload",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.StoreResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.StoreResponse": {
            "type": "object",
            "properties": {
                "gitoid": {
                    "type": "string"
                }
            }
        },
        "archivista.Resolver": {
            "type": "object"
        },
        "dsse.Envelope": {
            "type": "object",
            "properties": {
                "payload": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "payloadType": {
                    "type": "string"
                },
                "signatures": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dsse.Signature"
                    }
                }
            }
        },
        "dsse.Signature": {
            "type": "object",
            "properties": {
                "certificate": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "intermediates": {
                    "type": "array",
                    "items": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        }
                    }
                },
                "keyid": {
                    "type": "string"
                },
                "sig": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "timestamps": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dsse.SignatureTimestamp"
                    }
                }
            }
        },
        "dsse.SignatureTimestamp": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "type": {
                    "$ref": "#/definitions/dsse.SignatureTimestampType"
                }
            }
        },
        "dsse.SignatureTimestampType": {
            "type": "string",
            "enum": [
                "tsp"
            ],
            "x-enum-varnames": [
                "TimestampRFC3161"
            ]
        }
    }
}